+++ tests/test_lookup/tests.py
@@ -1191,17 +1191,18 @@ class LookupTests(DjangoLookupTests):
 
     def test_related_complement(self):
         tags = Tag.objects.filter(articles__author=OuterRef('id'), name='Tag 2')
-        self.assertQuerysetEqual(
-            Author.objects.filter(article__headline__startswith='Article', article__tag__complement=Exists(tags)).distinct(),
-            [
-                '<Author: Author object (1)>',
-                '<Author: Author object (2)>',
-            ],
+        self.assertEqual(
+            [a.name for a in Author.objects.filter(
+                article__headline__startswith='Article',
+                article__tag__complement=Exists(tags)).distinct()],
+            ['Author 1', 'Author 2'],
         )
 
     def test_negate_related_complement(self):
         tags = Tag.objects.filter(articles__author=OuterRef('id'), name='Tag 2')
-        self.assertQuerysetEqual(
-            Author.objects.filter(article__headline__startswith='Article', article__tag__complement=~Exists(tags)).distinct(),
+        self.assertEqual(
+            [a.name for a in Author.objects.filter(
+                article__headline__startswith='Article',
+                article__tag__complement=~Exists(tags)).distinct()],
             [],
         )