+++ lookup_extensions/related_lookups/__init__.py
@@ -1,0 +1,1 @@
+from .complements import *  # noqa F401,F403