+++ HISTORY.rst
@@ -5,7 +5,7 @@ History
 0.2.1
 -----
 
-- Support a related `complement` lookup
+- Support related `complement` lookups
 
 0.2.0
 -----